src/requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Deprecated==1.2.18
Flask==3.0.0
google-api-core==2.15.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-logging==3.11.4
google-cloud-ndb==2.2.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
html-sanitizer==2.2.0
idna==3.6
importlib_metadata==8.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.0.1
MarkupSafe==2.1.3
oauthlib==3.2.2
opentelemetry-api==1.29.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymemcache==4.0.0
pyOpenSSL==23.3.0
python-http-client==3.3.7
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.11.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
starkbank-ecdsa==2.2.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.17.2
zipp==3.21.0
